Deprecate older modules
